Move wifi files to their final location & cleanup
