+++ examples/platform/silabs/display/demo-ui.c
@@ -27,7 +27,7 @@
 #include "sl_memlcd.h"
 #include <app/icd/server/ICDServerConfig.h>
 #if SL_WIFI && !SLI_SI91X_MCU_INTERFACE
-#include <platform/silabs/wifi/wf200/platform/spi_multiplex.h>
+#include <platform/silabs/wifi/ncp-common/spi_multiplex.h>
 #endif // SL_WIFI && !SLI_SI91X_MCU_INTERFACE
 #include <string.h>
