+++ src/platform/silabs/efr32/OTAImageProcessorImpl.cpp
@@ -22,7 +22,7 @@
 #include <platform/silabs/SilabsConfig.h>
 
 #if SL_WIFI
-#include <platform/silabs/wifi/wf200/platform/spi_multiplex.h>
+#include <platform/silabs/wifi/ncp-common/spi_multiplex.h>
 #endif // SL_WIFI
 
 extern "C" {
