+++ src/platform/silabs/wifi/SiWx/ncp-platform/rs9117.gni
@@ -18,17 +18,14 @@ import("${silabs_sdk_build_root}/efr32_sdk.gni")
 
 rs911x_src_plat = [
   "${chip_root}/src/platform/silabs/wifi/SiWx/WifiInterface.cpp",
-  "${chip_root}/src/platform/silabs/wifi/rs911x/platform/rsi_hal_mcu_interrupt.c",
-  "${chip_root}/src/platform/silabs/wifi/rs911x/platform/sl_si91x_ncp_utility.c",
-  "${chip_root}/src/platform/silabs/wifi/rs911x/platform/sl_board_configuration.h",
+  "${chip_root}/src/platform/silabs/wifi/SiWx/ncp-platform/sl_si91x_ncp_utility.c",
+  "${chip_root}/src/platform/silabs/wifi/SiWx/ncp-platform/sl_board_configuration.h",
   "${chip_root}/src/platform/silabs/wifi/wiseconnect-abstraction/WiseconnectInterfaceAbstraction.cpp",
   "${chip_root}/src/platform/silabs/wifi/wiseconnect-abstraction/WiseconnectInterfaceAbstraction.h",
-
-  # TODO: We shouldn't need a file form the WF200 for the SiWx917 NCP builds
-  "${chip_root}/src/platform/silabs/wifi/wf200/platform/spi_multiplex.h",
+  "${chip_root}/src/platform/silabs/wifi/ncp-common/spi_multiplex.h",
 
   # TODO : We should be using the file from the Wiseconnect SDK and not our copy of it.
-  "${chip_root}/src/platform/silabs/wifi/rs911x/platform/efx32_ncp_host.c",
+  "${chip_root}/src/platform/silabs/wifi/SiWx/ncp-platform/efx32_ncp_host.c",
 ]
 
 rs9117_inc_plat = [ "${wifi_sdk_root}/components/si91x/ble/inc" ]
