+++ src/platform/silabs/wifi/SiWx/ncp-platform/sl_si91x_ncp_utility.c
@@ -15,9 +15,6 @@
  *    limitations under the License.
  */
 
-/**
- * Includes
- */
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
