+++ src/platform/silabs/wifi/rs911x/rs911x.gni
@@ -18,12 +18,12 @@ import("${silabs_sdk_build_root}/efr32_sdk.gni")
 
 rs911x_src_plat = [
   "${chip_root}/src/platform/silabs/wifi/rs911x/WifiInterface.cpp",
-  "${chip_root}/src/platform/silabs/wifi/rs911x/platform/rsi_hal_mcu_interrupt.c",
-  "${chip_root}/src/platform/silabs/wifi/rs911x/platform/rsi_hal_mcu_ioports.c",
-  "${chip_root}/src/platform/silabs/wifi/rs911x/platform/rsi_hal_mcu_timer.c",
-  "${chip_root}/src/platform/silabs/wifi/rs911x/platform/efx_spi.c",
-  "${chip_root}/src/platform/silabs/wifi/rs911x/platform/rsi_board_configuration.h",
-  "${chip_root}/src/platform/silabs/wifi/rs911x/platform/rsi_hal.h",
+  "${chip_root}/src/platform/silabs/wifi/rs911x/ncp-platform/rsi_hal_mcu_interrupt.c",
+  "${chip_root}/src/platform/silabs/wifi/rs911x/ncp-platform/rsi_hal_mcu_ioports.c",
+  "${chip_root}/src/platform/silabs/wifi/rs911x/ncp-platform/rsi_hal_mcu_timer.c",
+  "${chip_root}/src/platform/silabs/wifi/rs911x/ncp-platform/efx_spi.c",
+  "${chip_root}/src/platform/silabs/wifi/rs911x/ncp-platform/rsi_board_configuration.h",
+  "${chip_root}/src/platform/silabs/wifi/rs911x/ncp-platform/rsi_hal.h",
   "${chip_root}/src/platform/silabs/wifi/wiseconnect-abstraction/WiseconnectInterfaceAbstraction.cpp",
 ]
