+++ src/platform/silabs/wifi/wf200/wf200.gni
@@ -18,12 +18,12 @@ import("${silabs_sdk_build_root}/efr32_sdk.gni")
 
 wf200_plat_src = [
   "${chip_root}/src/platform/silabs/wifi/wf200/WifiInterface.cpp",
-  "${chip_root}/src/platform/silabs/wifi/wf200/platform/sl_wfx_task.c",
-  "${chip_root}/src/platform/silabs/wifi/wf200/platform/wf200_init.c",
-  "${chip_root}/src/platform/silabs/wifi/wf200/platform/efr_spi.c",
-  "${chip_root}/src/platform/silabs/wifi/wf200/platform/sl_wfx_board.h",
-  "${chip_root}/src/platform/silabs/wifi/wf200/platform/sl_wfx_host.h",
-  "${chip_root}/src/platform/silabs/wifi/wf200/platform/sl_wfx_task.h",
-  "${chip_root}/src/platform/silabs/wifi/wf200/platform/sl_custom_board.h",
-  "${chip_root}/src/platform/silabs/wifi/wf200/platform/spi_multiplex.h",
+  "${chip_root}/src/platform/silabs/wifi/wf200/ncp-platform/sl_wfx_task.c",
+  "${chip_root}/src/platform/silabs/wifi/wf200/ncp-platform/wf200_init.c",
+  "${chip_root}/src/platform/silabs/wifi/wf200/ncp-platform/efr_spi.c",
+  "${chip_root}/src/platform/silabs/wifi/wf200/ncp-platform/sl_wfx_board.h",
+  "${chip_root}/src/platform/silabs/wifi/wf200/ncp-platform/sl_wfx_host.h",
+  "${chip_root}/src/platform/silabs/wifi/wf200/ncp-platform/sl_wfx_task.h",
+  "${chip_root}/src/platform/silabs/wifi/wf200/ncp-platform/sl_custom_board.h",
+  "${chip_root}/src/platform/silabs/wifi/ncp-common/spi_multiplex.h",
 ]
